requirements.txt
python-dotenv == 1.0.0
uvicorn ==0.23.2
ruff ==0.3.2
fastapi ==0.110.0
flask ==3.0.2
flask-sock ==0.7.0
assemblyai ==0.23.0
twilio ==9.0.0
urllib3 ==2.2.1
openai ==1.55.3
vonage ==3.13.0
python-multipart==0.0.9
redis==5.0.3
retell-sdk==4.19.0
python-dateutil==2.9.0.post0
firebase-admin==6.5.0
stripe==10.4.0
boto3==1.37.1
apscheduler==3.11.0